Move ExceptionFactory from internal package
