+++ plugin/src/main/groovy/com/novoda/buildproperties/internal/MapEntries.groovy
@@ -2,6 +2,7 @@ package com.novoda.buildproperties.internal
 
 import com.novoda.buildproperties.Entries
 import com.novoda.buildproperties.Entry
+import com.novoda.buildproperties.ExceptionFactory
 
 class MapEntries implements Entries {
     private final Map<String, Object> map
